+++ Adjust/sdk-core/src/main/java/com/adjust/sdk/AdjustTimeoutCallback.java
@@ -5,6 +5,7 @@ import com.adjust.sdk.scheduler.TimerOnce;
 public class AdjustTimeoutCallback {
     private OnAdidReadListener onAdidReadListener;
     private OnAttributionReadListener onAttributionReadListener;
+    private OnThirdPartySharingSettingsReadListener onThirdPartySharingSettingsReadListener;
     private TimerOnce timeoutTimer;
 
     public AdjustTimeoutCallback(OnAdidReadListener onAdidReadListener) {
@@ -15,12 +16,20 @@ public class AdjustTimeoutCallback {
         this.onAttributionReadListener = onAttributionReadListener;
     }
 
+    public AdjustTimeoutCallback(OnThirdPartySharingSettingsReadListener onThirdPartySharingSettingsReadListener) {
+        this.onThirdPartySharingSettingsReadListener = onThirdPartySharingSettingsReadListener;
+    }
+
     public void setOnAdidReadListener(OnAdidReadListener onAdidReadListener) {
         this.onAdidReadListener = onAdidReadListener;
     }
 
     public void setOnAttributionReadListener(OnAttributionReadListener onAttributionReadListener) {
         this.onAttributionReadListener = onAttributionReadListener;
+    }
+
+    public void setOnThirdPartySharingSettingsReadListener(OnThirdPartySharingSettingsReadListener onThirdPartySharingSettingsReadListener) {
+        this.onThirdPartySharingSettingsReadListener = onThirdPartySharingSettingsReadListener;
     }
 
     public void setTimer(TimerOnce timer) {
@@ -33,6 +42,10 @@ public class AdjustTimeoutCallback {
 
     public OnAttributionReadListener getOnAttributionReadListener() {
         return onAttributionReadListener;
+    }
+
+    public OnThirdPartySharingSettingsReadListener getOnThirdPartySharingSettingsReadListener() {
+        return onThirdPartySharingSettingsReadListener;
     }
 
     public TimerOnce getTimeoutTimer() {