+++ Adjust/sdk-core/src/main/java/com/adjust/sdk/IActivityHandler.java
@@ -101,6 +101,8 @@ public interface IActivityHandler {
 
     void getAttributionWithTimeout(long timeoutInMilliSec, OnAttributionReadListener onAttributionReadListener);
 
+    void getThirdPartySharingSettingsWithTimeout(long timeoutInMilliSec, OnThirdPartySharingSettingsReadListener onThirdPartySharingSettingsReadListener);
+
     AdjustConfig getAdjustConfig();
 
     DeviceInfo getDeviceInfo();