+++ Adjust/sdk-core/src/main/java/com/adjust/sdk/OnThirdPartySharingSettingsReadListener.java
@@ -1,0 +1,5 @@
+package com.adjust.sdk;
+
+public interface OnThirdPartySharingSettingsReadListener {
+    void onThirdPartySharingSettingsRead(AdjustThirdPartySharingResult adjustThirdPartySharingResult);
+}