+++ Adjust/tests/test-app-core/src/main/java/com/adjust/testapp/AdjustCommandExecutor.java
@@ -27,8 +27,8 @@ import com.adjust.sdk.AdjustSessionSuccess;
 import com.adjust.sdk.AdjustStoreInfo;
 import com.adjust.sdk.AdjustTestOptions;
 import com.adjust.sdk.AdjustThirdPartySharing;
+import com.adjust.sdk.AdjustThirdPartySharingResult;
 import com.adjust.sdk.LogLevel;
-import com.adjust.sdk.OnAdidReadListener;
 import com.adjust.sdk.OnAttributionChangedListener;
 import com.adjust.sdk.OnDeeplinkResolvedListener;
 import com.adjust.sdk.OnDeferredDeeplinkResponseListener;
@@ -39,6 +39,7 @@ import com.adjust.sdk.OnPurchaseVerificationFinishedListener;
 import com.adjust.sdk.OnRemoteTriggerListener;
 import com.adjust.sdk.OnSessionTrackingFailedListener;
 import com.adjust.sdk.OnSessionTrackingSucceededListener;
+import com.adjust.sdk.OnThirdPartySharingSettingsChangedListener;
 import com.adjust.test_options.TestConnectionOptions;
 
 import java.util.HashMap;
@@ -109,6 +110,7 @@ public class AdjustCommandExecutor {
                 case "sdkVersionGetter" : sdkVersionGetter(); break;
                 case "googleAdIdGetter" : googleAdIdGetter(); break;
                 case "amazonAdIdGetter" : amazonAdIdGetter(); break;
+                case "tpsSettingsGetter" : tpsSettingsGetter(); break;
                 //case "testBegin": testBegin(); break;
                 // case "testEnd": testEnd(); break;
             }
@@ -397,6 +399,21 @@ public class AdjustCommandExecutor {
             });
         }
 
+        if (command.containsParameter("thirdPartySharingSettingsChangedCallbackSendAll")) {
+            final String localBasePath = basePath;
+            adjustConfig.setOnThirdPartySharingSettingsChangedListener(new OnThirdPartySharingSettingsChangedListener() {
+                @Override
+                public void onThirdPartySharingSettingsChanged(AdjustThirdPartySharingResult result) {
+                    Map<String, String> fields = new HashMap<>();
+                    if (result != null && result.getThirdPartySharingSettingsJson() != null) {
+                        fields.put("third_party_sharing_settings", result.getThirdPartySharingSettingsJson());
+                    }
+                    MainActivity.testLibrary.setInfoToSend(fields);
+                    MainActivity.testLibrary.sendInfoToServer(localBasePath);
+                }
+            });
+        }
+
         if (command.containsParameter("sessionCallbackSendSuccess")) {
             final String localBasePath = basePath;
             adjustConfig.setOnSessionTrackingSucceededListener(new OnSessionTrackingSucceededListener() {
@@ -525,6 +542,14 @@ public class AdjustCommandExecutor {
             boolean appSetIdReadingEnabled = "true".equals(appSetIdReadingEnabledS);
             if (!appSetIdReadingEnabled) {
                 adjustConfig.disableAppSetIdReading();
+            }
+        }
+
+        if (command.containsParameter("fbIdReadingEnabled")) {
+            String fbIdReadingEnabledS = command.getFirstParameterValue("fbIdReadingEnabled");
+            boolean fbIdReadingEnabled = "true".equals(fbIdReadingEnabledS);
+            if (!fbIdReadingEnabled) {
+                adjustConfig.disableFbIdReading();
             }
         }
     }
@@ -1022,6 +1047,21 @@ public class AdjustCommandExecutor {
             } else {
                 MainActivity.testLibrary.addInfoToSend("adid", "null");
             }
+            MainActivity.testLibrary.addInfoToSend("test_callback_id", testCallbackId);
+            MainActivity.testLibrary.sendInfoToServer(basePath);
+        });
+    }
+
+    private void tpsSettingsGetter() {
+        long timeout = Long.parseLong(command.getFirstParameterValue("timeout"));
+        String testCallbackId = command.getFirstParameterValue("testCallbackId");
+
+        Adjust.getThirdPartySharingSettingsWithTimeout(context, timeout, adjustThirdPartySharingResult -> {
+            Map<String, String> fields = new HashMap<>();
+            if (adjustThirdPartySharingResult != null) {
+                fields.put("third_party_sharing", adjustThirdPartySharingResult.getThirdPartySharingSettingsJson());
+            }
+            MainActivity.testLibrary.setInfoToSend(fields);
             MainActivity.testLibrary.addInfoToSend("test_callback_id", testCallbackId);
             MainActivity.testLibrary.sendInfoToServer(basePath);
         });